refactor: provider 폴더 계층 변경
