+++ lib/presentation/screen/home/home_screen.dart
@@ -1,6 +1,5 @@
 import 'package:calendar_scheduler/presentation/const/strings.dart';
 import 'package:calendar_scheduler/presentation/extension.dart';
-import 'package:calendar_scheduler/presentation/provider/schedule_provider.dart';
 import 'package:calendar_scheduler/presentation/screen/component/app_bar.dart';
 import 'package:calendar_scheduler/presentation/screen/component/calendar.dart';
 import 'package:calendar_scheduler/presentation/screen/component/date_banner.dart';
@@ -13,6 +12,7 @@ import 'package:provider/provider.dart';
 
 import '../../const/colors.dart';
 import '../component/schedule_register_bottom_sheet.dart';
+import '../provider/schedule_provider.dart';
 
 final _FIRST_DAY = DateTime.utc(1900);
 final _LAST_DAY = DateTime.utc(3000);
